project/requirements.txt
fastapi==0.68.1
alembic==1.7.1
psycopg2-binary==2.8.6
fastapi-sqlalchemy==0.2.1
uvicorn==0.15.0
python-dotenv==0.20.0